src/dal/requirements.txt
requests~=2.27.1
setuptools~=60.2.0
pydantic==2.4.2
fastapi_offline
uvicorn==0.17.6
broadcaster
asyncio_redis
logbook
websockets
fastapi-login
werkzeug
environs
faker
sentry-sdk[pymongo]
pytz
decorator
aiohttp
motor
ariadne
pandas==1.2.3
pyodbc
